Basic fixes and updates
Nicer syntax, simplified OTFCommand to one function in RobotContainer
